# fastapi_app/openapi_baseline.json
{
  "components": {
    "schemas": {
...
      "RequestValidationErrorPayload": {
        "description": "Field-preserving Pydantic request validation error.",
        "properties": {
          "code": {
            "const": "REQUEST_VALIDATION_ERROR",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "RequestValidationErrorPayload",
        "type": "object"
      },
      "RequestValidationErrorResponse": {
        "description": "Documented 422 form of the maintained response envelope.",
        "properties": {
          "data": {
            "title": "Data",
            "type": "null"
          },
          "error": {
            "$ref": "#/components/schemas/RequestValidationErrorPayload"
          },
          "success": {
            "const": false,
            "default": false,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "error"
        ],
        "title": "RequestValidationErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Structural Engineering Library",
      "url": "https://github.com/yourusername/structural_engineering_lib"
    },
    "description": "\n## IS 456:2000 Compliant Structural Engineering Library\n\nThis API exposes the library's route-specific supported IS 456:2000 reinforced\nconcrete calculations. Each result remains subject to the documented case\nboundary and qualified engineering review.\n\n### Features\n\n- **Beam Design**: Flexure, shear, and combined design calculations\n- **Column Design**: Bounded rectangular short/long-column workflows\n- **Footing Design**: Isolated-footing checks and concentric load transfer\n- **Slab Design**: Bounded simply supported one-way slab strip\n- **Detailing**: Reinforcement layout, spacing, and development lengths\n- **Optimization**: Cost-optimized beam cross-section selection\n- **Smart Analysis**: AI-assisted design suggestions and insights\n- **3D Geometry**: Visualization-ready mesh generation\n\n### Design Codes\n\n- IS 456:2000 - Plain and Reinforced Concrete\n- IS 13920 - Ductile Detailing for Seismic Resistance\n\n### Units\n\nAll inputs and outputs use consistent units:\n- **Length**: millimeters (mm)\n- **Force**: kilonewtons (kN)\n- **Moment**: kilonewton-meters (kN\u00b7m)\n- **Stress**: N/mm\u00b2 (MPa)\n- **Area**: mm\u00b2\n",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Structural Engineering API",
...
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/stream/batch-design": {
      "get": {
        "description": "Stream batch beam design results.\n\nAccepts JSON array of beam parameters and streams results as they complete.\n\nEvents:\n    - start: Job started with job_id and total count\n    - progress: Current progress (completed/total)\n    - design_result: Individual beam result\n    - error: Individual beam error\n    - complete: All beams processed\n\nExample:\n    ```javascript\n    const beams = JSON.stringify([\n        {width: 300, depth: 500, moment: 100, fck: 25, fy: 500},\n        {width: 300, depth: 500, moment: 150, fck: 25, fy: 500}\n    ]);\n    const es = new EventSource(`/stream/batch-design?beams=${encodeURIComponent(beams)}`);\n\n    es.addEventListener('design_result', (e) => {\n        const result = JSON.parse(e.data);\n        console.log('Beam completed:', result);\n    });\n\n    es.addEventListener('complete', (e) => {\n        console.log('All done!');\n        es.close();\n    });\n    ```",
        "operationId": "stream_batch_design_stream_batch_design_get",
        "parameters": [
          {
            "description": "JSON array of beam parameters",
            "in": "query",
            "name": "beams",
            "required": true,
            "schema": {
              "description": "JSON array of beam parameters",
              "title": "Beams",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestValidationErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Stream Batch Design",
        "tags": [
          "streaming"
        ]
      }
    },
    "/stream/job/{job_id}": {
      "get": {
        "description": "Get status of a batch job.\n\nReturns job progress, results count, and any errors.",
        "operationId": "get_job_status_stream_job__job_id__get",
        "parameters": [
          {
            "description": "Batch job ID",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "description": "Batch job ID",
              "pattern": "^[a-f0-9]{8}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Status Stream Job  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestValidationErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Job Status",
        "tags": [
          "streaming"
        ]
      }
    }
  },
...
}